# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Authify
  description: 'Authify Description'
  version: 1.0.0
servers:
  -
    url: 'http://authify.test'
paths:
...
  /api/v1/set-new-password:
    post:
      summary: 'Set New Password'
      operationId: setNewPassword
      description: 'Atur ulang kata sandi pengguna menggunakan kode verifikasi.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Password reset successfully'
                properties:
                  message:
                    type: string
                    example: 'Password reset successfully'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Invalid or expired verification code'
                properties:
                  message:
                    type: string
                    example: 'Invalid or expired verification code'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  description: 'Nomor telepon pengguna.'
                  example: '+1234567890'
                  nullable: false
                code:
                  type: string
                  description: 'Kode verifikasi 6 digit.'
                  example: '123456'
                  nullable: false
                password:
                  type: string
                  description: 'Kata sandi baru.'
                  example: newpassword123
                  nullable: false
                password_confirmation:
                  type: string
                  description: 'Konfirmasi kata sandi baru.'
                  example: newpassword123
                  nullable: false
              required:
                - phone
                - code
                - password
                - password_confirmation
      security: []
...